+++ home/sheath.nix
@@ -34,7 +34,7 @@ in
     inputs.sops-nix.homeManagerModules.sops
     inputs.nix-index-database.homeModules.nix-index
   ] ++ lib.optionals workstation [
-    ./ghostty.nix
+    ./alacritty.nix
     ./ssh.nix
     ./vscode.nix
     ./monitors.nix